fixing unsigned ints and warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@22332 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSTrafficLightLogic.cpp
@@ -289,7 +289,7 @@ MSTrafficLightLogic::setCurrentDurationIncrement(SUMOTime delay) {
 void MSTrafficLightLogic::initMesoTLSPenalties() {
     // set mesoscopic time penalties
     const Phases& phases = getPhases();
-    const int numLinks = myLinks.size();
+    const int numLinks = (int)myLinks.size();
     // warning already given if not all states are used
     assert(numLinks <= (int)phases.front()->getState().size());
     SUMOTime duration = 0;
